requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.27.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0
apscheduler==3.10.4
python-telegram-bot==21.5
lxml==5.3.0
